Add initial README.md
Added an initial version of README.md to my GitHub repository. The README provides an overview of my skills, interests, and background as a second-year Bachelor's of Engineering student. It highlights my proficiency in C, C++, Java, Python, and front-end development, as well as my current focus on learning Artificial Intelligence and Machine Learning. Additionally, it includes a brief introduction about myself and invites others to connect with me on LinkedIn and Twitter. Looking forward to further enhancing my GitHub profile and engaging with the developer community.